+++ server/src/main.cpp
@@ -4,11 +4,11 @@
 
 using namespace std;
 
-sock_info_t         g_sock_info;
-map<string,int>     g_name_sock;
-map<int,string>     g_sock_name;
-fd_set              g_all_set;
-int                 g_max_fd(0);
+sock_info_t         	g_sock_info;
+map<string,int16_t>     g_name_sock;
+map<int16_t,string>     g_sock_name;
+fd_set              	g_all_set;
+int16_t                 g_max_fd(0);
 
 int main(int argc, char* argv[])
 {
@@ -36,7 +36,7 @@ int main(int argc, char* argv[])
 			default:
 				if(g_sock_name.size() > 0)
 				{
-					for(map<int,string>::iterator it = g_sock_name.begin(); it != g_sock_name.end(); it++)
+					for(map<int16_t,string>::iterator it = g_sock_name.begin(); it != g_sock_name.end(); it++)
 					{
 						if(FD_ISSET((*it).first, &cur_set))
 						{
@@ -46,7 +46,7 @@ int main(int argc, char* argv[])
 				}
 				if(FD_ISSET(g_sock_info.get_ser_sock(), &cur_set))
 				{
-					int conn_fd = g_sock_info.accept_cli();
+					int16_t conn_fd = g_sock_info.accept_cli();
 					g_sock_name[conn_fd] = "";
 					FD_SET(conn_fd, &g_all_set);
 					g_max_fd = g_max_fd < conn_fd ? conn_fd:g_max_fd;
